+++ plugins/search-plugin/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/lib.routes';
+
+export * from './lib/search-plugin/search-plugin.component';