+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- New `Search Plugin` with minimalistic search experience
+- Support icon buttons with Header component
+
 ## [3.0.0] - 2024-03-12
 
 ### Added